Move suggestions to parent.
